I rename the modal components
