+++ dae-alright/src/components/Modal/CustomModal/index.scss
