+++ dae-alright/src/components/Modal/SignUpModal/index.scss
